backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
groq>=0.11.0
supabase>=2.8.0
python-dotenv>=1.0.1
pydantic>=2.8.0
